+++ src/com/twitter/intellij/pants/settings/ImportFromPantsControl.java
@@ -1,15 +1,11 @@
-// Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
+// Copyright 2017 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-package com.twitter.intellij.pants.service.settings;
+package com.twitter.intellij.pants.settings;
 
 import com.intellij.openapi.externalSystem.service.settings.AbstractImportFromExternalSystemControl;
 import com.intellij.openapi.externalSystem.util.ExternalSystemSettingsControl;
 import com.intellij.openapi.util.text.StringUtil;
-import com.twitter.intellij.pants.settings.PantsProjectSettings;
-import com.twitter.intellij.pants.settings.PantsProjectSettingsControl;
-import com.twitter.intellij.pants.settings.PantsSettings;
-import com.twitter.intellij.pants.settings.PantsSettingsListener;
 import com.twitter.intellij.pants.util.PantsConstants;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
